requirements.txt
argcomplete
getmac
pyjwt
netaddr
Pyro4
psutil
ruamel.yaml
sqlitedict
pyserial
vpp-papi
websocket-client
pyroute2
